Refactor: restructured code
