+++ packages/ledger/.img/ledger_test_SetBranchConfirmed.png
