+++ packages/ledger/vm/devnetvm/outputtype.go
